+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 * Add the `canu.show_login_page` method.
 * Add the `canu.show_profile_page` method.
 * Add the `canu.set_page` method.
+* Remove the `canu.logout` method.
 
 ## 0.3.0 (2024-06-04)
 * Add the `canu.functions.retrieve_from_web` method.