+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # CHANGELOG
 
-## 0.4.0 (in development)
+## 0.4.0 (2024-06-06)
 * Rename the `canu.get_authenticator` method to `canu.authenticate`.
 * Add the `canu.update_yaml_file` method.
 * Add the `canu.show_login_page` method.